Remove extra inheritance
