+++ src/main/kotlin/io/emeraldpay/dshackle/upstream/ethereum/EthereumLowerBoundStateDetector.kt
@@ -3,15 +3,19 @@ package io.emeraldpay.dshackle.upstream.ethereum
 import io.emeraldpay.dshackle.upstream.ChainRequest
 import io.emeraldpay.dshackle.upstream.ChainResponse
 import io.emeraldpay.dshackle.upstream.Upstream
+import io.emeraldpay.dshackle.upstream.lowerbound.LowerBoundData
+import io.emeraldpay.dshackle.upstream.lowerbound.LowerBoundDetector
 import io.emeraldpay.dshackle.upstream.lowerbound.LowerBoundType
-import io.emeraldpay.dshackle.upstream.lowerbound.detector.RecursiveLowerBoundDetector
+import io.emeraldpay.dshackle.upstream.lowerbound.detector.RecursiveLowerBound
 import io.emeraldpay.dshackle.upstream.lowerbound.toHex
 import io.emeraldpay.dshackle.upstream.rpcclient.ListParams
+import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
 
 class EthereumLowerBoundStateDetector(
     private val upstream: Upstream,
-) : RecursiveLowerBoundDetector(upstream) {
+) : LowerBoundDetector() {
+    private val recursiveLowerBound = RecursiveLowerBound(upstream, LowerBoundType.STATE, nonRetryableErrors)
 
     companion object {
         private val nonRetryableErrors = setOf(
@@ -40,23 +44,22 @@ class EthereumLowerBoundStateDetector(
         )
     }
 
-    override fun hasData(block: Long): Mono<Boolean> {
-        if (block == 0L) {
-            return Mono.just(true)
-        }
-        return upstream.getIngressReader().read(
-            ChainRequest(
-                "eth_getBalance",
-                ListParams(ZERO_ADDRESS, block.toHex()),
-            ),
-        )
-            .retryWhen(retrySpec(nonRetryableErrors))
-            .flatMap(ChainResponse::requireResult)
-            .map { true }
-            .onErrorReturn(false)
+    override fun period(): Long {
+        return 5
     }
 
-    override fun type(): LowerBoundType {
-        return LowerBoundType.STATE
+    override fun internalDetectLowerBound(): Flux<LowerBoundData> {
+        return recursiveLowerBound.recursiveDetectLowerBound { block ->
+            if (block == 0L) {
+                Mono.just(ChainResponse(ByteArray(0), null))
+            } else {
+                upstream.getIngressReader().read(
+                    ChainRequest(
+                        "eth_getBalance",
+                        ListParams(ZERO_ADDRESS, block.toHex()),
+                    ),
+                )
+            }
+        }
     }
 }
